requirements.txt
boto
dnspython
